langchain_service/requirements.txt
flask==3.0.2
requests==2.32.3
python-dotenv==1.0.1
matplotlib==3.8.4
